clients/web/requirements.txt
streamlit>=1.36.0
requests>=2.31.0
Pillow>=10.0.0
